refactor notification section
